Moved memcached handling to helper module
